Add getInteractionStatusFromEvent to msal-browser, update react and angular to use function
