+++ lib/msal-browser/src/index.ts
@@ -5,7 +5,7 @@
 
 export { PublicClientApplication } from "./app/PublicClientApplication";
 export { Configuration } from "./config/Configuration";
-export { InteractionType, BrowserCacheLocation } from "./utils/BrowserConstants";
+export { InteractionType, InteractionStatus, BrowserCacheLocation } from "./utils/BrowserConstants";
 export { BrowserUtils } from "./utils/BrowserUtils";
 
 // Browser Errors
@@ -22,7 +22,7 @@ export { EndSessionRequest } from "./request/EndSessionRequest";
 export { AuthorizationUrlRequest } from "./request/AuthorizationUrlRequest";
 
 // Events
-export { EventMessage, EventPayload, EventError, EventCallbackFunction } from "./event/EventMessage";
+export { EventMessage, EventPayload, EventError, EventCallbackFunction, EventMessageUtils } from "./event/EventMessage";
 export { EventType } from "./event/EventType";
 
 // Common Object Formats
