+++ lib/msal-react/src/MsalProvider.tsx
@@ -7,12 +7,13 @@ import React, { useState, useEffect, PropsWithChildren, useMemo } from "react";
 import {
     IPublicClientApplication,
     EventType,
-    EventMessage, 
-    InteractionType,
+    EventMessage,
+    EventMessageUtils,
+    InteractionStatus,
     Logger
 } from "@azure/msal-browser";
 import { MsalContext, IMsalContext } from "./MsalContext";
-import { Constants, InteractionStatus } from "./utils/Constants";
+import { Constants } from "./utils/Constants";
 import { accountArraysAreEqual } from "./utils/utilities";
 import { AccountIdentifiers } from "./types/AccountIdentifiers";
 
@@ -65,45 +66,10 @@ export function MsalProvider({instance, children}: MsalProviderProps): React.Rea
 
     useEffect(() => {
         const callbackId = instance.addEventCallback((message: EventMessage) => {
-            switch (message.eventType) {
-                case EventType.LOGIN_START:
-                    logger.info("MsalProvider - Login called, setting inProgress to 'login'");
-                    setInProgress(InteractionStatus.Login);
-                    break;
-                case EventType.SSO_SILENT_START:
-                    logger.info("MsalProvider - SsoSilent called, setting inProgress to 'ssoSilent'");
-                    setInProgress(InteractionStatus.SsoSilent);
-                    break;
-                case EventType.ACQUIRE_TOKEN_START:
-                    if (message.interactionType === InteractionType.Redirect || message.interactionType === InteractionType.Popup) {
-                        logger.info("MsalProvider - Interactive acquireToken called, setting inProgress to 'acquireToken'");
-                        setInProgress(InteractionStatus.AcquireToken);
-                    }
-                    break;
-                case EventType.HANDLE_REDIRECT_START:
-                    logger.info("MsalProvider - HandleRedirectPromise called, setting inProgress to 'handleRedirect'");
-                    setInProgress(InteractionStatus.HandleRedirect);
-                    break;
-                case EventType.LOGOUT_START:
-                    logger.info("MsalProvider - Logout called, setting inProgress to 'logout'");
-                    setInProgress(InteractionStatus.Logout);
-                    break;
-                case EventType.LOGIN_SUCCESS:
-                case EventType.SSO_SILENT_SUCCESS:
-                case EventType.HANDLE_REDIRECT_END:
-                case EventType.LOGIN_FAILURE:
-                case EventType.SSO_SILENT_FAILURE:
-                case EventType.LOGOUT_FAILURE:
-                    logger.info("MsalProvider - Interactive request finished, setting inProgress to 'none'");
-                    setInProgress(InteractionStatus.None);
-                    break;
-                case EventType.ACQUIRE_TOKEN_SUCCESS:
-                case EventType.ACQUIRE_TOKEN_FAILURE:
-                    if (message.interactionType === InteractionType.Redirect || message.interactionType === InteractionType.Popup) {
-                        logger.info("MsalProvider - Interactive acquireToken request finished, setting inProgress to 'none'");
-                        setInProgress(InteractionStatus.None);
-                    }
-                    break;
+            const status = EventMessageUtils.getInteractionStatusFromEvent(message);
+            if (status !== null) {
+                logger.info(`MsalProvider - setting inProgress to ${status}`);
+                setInProgress(status);
             }
         });
         logger.verbose(`MsalProvider - Registered event callback with id: ${callbackId}`);
