+++ lib/msal-react/src/utils/Constants.ts
@@ -7,13 +7,3 @@ export const Constants = {
     SKU: "@azure/msal-react",
     VERSION: "1.0.0-alpha.3"
 };
-
-export enum InteractionStatus {
-    Startup = "startup",
-    Login = "login",
-    Logout = "logout",
-    AcquireToken = "acquireToken",
-    SsoSilent = "ssoSilent",
-    HandleRedirect = "handleRedirect",
-    None = "none"
-}
